Update how-to-enroll.md, remove same-site explainer link
The explanation for where attestations json file should be placed is confusing.

Namely the link we propose to be removed leads to a page where "same-site" is explained but the file cannot be placed on the same site as we were notified during the enrollment process.

If you enroll the site: example.com, the file should be placed at
example.com/.well-known/privacy-sanbox-attestations.json.

Putting it on a same-site subdomain will fail the automatic tests. For example: privacy-sandbox.example.com/.well-known/privacy-sanbox-attestations.json.